# test/e2e/__snapshots__/test_openapi/test_OpenAPIの形が変わっていないことを確認.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UserDictWord": {
        "description": "ユーザー辞書のビルドに必要な単語情報\n単語登録・変更リクエストで受け取った単語情報のバリデーションと JSON への保存に用いる",
        "properties": {
          "accent_associative_rule": {
            "description": "アクセント結合規則",
            "title": "Accent Associative Rule",
            "type": "string"
          },
          "accent_type": {
            "description": "アクセント型",
            "items": {
              "type": "integer"
            },
            "title": "Accent Type",
            "type": "array"
          },
          "context_id": {
            "default": 1348,
            "description": "文脈 ID",
            "title": "Context Id",
            "type": "integer"
          },
          "inflectional_form": {
            "description": "活用形",
            "title": "Inflectional Form",
            "type": "string"
          },
          "inflectional_type": {
            "description": "活用型",
            "title": "Inflectional Type",
            "type": "string"
          },
          "mora_count": {
            "description": "モーラ数",
            "items": {
              "type": "integer"
            },
            "title": "Mora Count",
            "type": "array"
          },
          "part_of_speech": {
            "description": "品詞",
            "title": "Part Of Speech",
            "type": "string"
          },
          "part_of_speech_detail_1": {
            "description": "品詞細分類1",
            "title": "Part Of Speech Detail 1",
            "type": "string"
          },
          "part_of_speech_detail_2": {
            "description": "品詞細分類2",
            "title": "Part Of Speech Detail 2",
            "type": "string"
          },
          "part_of_speech_detail_3": {
            "description": "品詞細分類3",
            "title": "Part Of Speech Detail 3",
            "type": "string"
          },
          "priority": {
            "description": "優先度",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Priority",
            "type": "integer"
          },
          "pronunciation": {
            "description": "発音",
            "items": {
              "type": "string"
            },
            "title": "Pronunciation",
            "type": "array"
          },
          "stem": {
            "description": "原形",
            "items": {
              "type": "string"
            },
            "title": "Stem",
            "type": "array"
          },
          "surface": {
            "description": "表層形",
            "title": "Surface",
            "type": "string"
          },
          "yomi": {
            "description": "読み",
            "items": {
              "type": "string"
            },
            "title": "Yomi",
            "type": "array"
          }
        },
        "required": [
          "surface",
          "priority",
          "part_of_speech",
          "part_of_speech_detail_1",
          "part_of_speech_detail_2",
          "part_of_speech_detail_3",
          "inflectional_type",
          "inflectional_form",
          "stem",
          "yomi",
          "pronunciation",
          "accent_type",
          "accent_associative_rule"
        ],
        "title": "UserDictWord",
        "type": "object"
      },
      "UserDictWordForCompat": {
        "description": "UserDictWord とほとんど同じだが、ユーザー辞書関連 API の後方互換性を保つための互換レイヤー\nstem, yomi, pronunciation, accent_type, mora_count はリストではなく文字列/数値で表す\n既に UserDictWord にバリデーションが実装されていることから、重複するバリデーション定義は削られている",
        "properties": {
          "accent_associative_rule": {
            "description": "アクセント結合規則",
            "title": "Accent Associative Rule",
            "type": "string"
          },
          "accent_type": {
            "description": "アクセント型",
            "title": "Accent Type",
            "type": "integer"
          },
          "context_id": {
            "default": 1348,
            "description": "文脈 ID",
            "title": "Context Id",
            "type": "integer"
          },
          "inflectional_form": {
            "description": "活用形",
            "title": "Inflectional Form",
            "type": "string"
          },
          "inflectional_type": {
            "description": "活用型",
            "title": "Inflectional Type",
            "type": "string"
          },
          "mora_count": {
            "description": "モーラ数",
            "title": "Mora Count",
            "type": "integer"
          },
          "part_of_speech": {
            "description": "品詞",
            "title": "Part Of Speech",
            "type": "string"
          },
          "part_of_speech_detail_1": {
            "description": "品詞細分類1",
            "title": "Part Of Speech Detail 1",
            "type": "string"
          },
          "part_of_speech_detail_2": {
            "description": "品詞細分類2",
            "title": "Part Of Speech Detail 2",
            "type": "string"
          },
          "part_of_speech_detail_3": {
            "description": "品詞細分類3",
            "title": "Part Of Speech Detail 3",
            "type": "string"
          },
          "priority": {
            "description": "優先度",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Priority",
            "type": "integer"
          },
          "pronunciation": {
            "description": "発音",
            "title": "Pronunciation",
            "type": "string"
          },
          "stem": {
            "description": "原形",
            "title": "Stem",
            "type": "string"
          },
          "surface": {
            "description": "表層形",
            "title": "Surface",
            "type": "string"
          },
          "yomi": {
            "description": "読み",
            "title": "Yomi",
            "type": "string"
          }
        },
        "required": [
          "surface",
          "priority",
          "part_of_speech",
          "part_of_speech_detail_1",
          "part_of_speech_detail_2",
          "part_of_speech_detail_3",
          "inflectional_type",
          "inflectional_form",
          "stem",
          "yomi",
          "pronunciation",
          "accent_type",
          "accent_associative_rule"
        ],
        "title": "UserDictWordForCompat",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WordTypes": {
        "description": "品詞を表す列挙型",
        "enum": [
          "PROPER_NOUN",
          "LOCATION_NAME",
          "ORGANIZATION_NAME",
          "PERSON_NAME",
          "PERSON_FAMILY_NAME",
          "PERSON_GIVEN_NAME",
          "COMMON_NOUN",
          "VERB",
          "ADJECTIVE",
          "SUFFIX"
        ],
        "title": "WordTypes",
        "type": "string"
      }
    }
  },
  "info": {
    "description": "AivisSpeech の音声合成エンジンです。",
    "title": "AivisSpeech Engine",
    "version": "latest"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/import_user_dict": {
      "post": {
        "description": "他のユーザー辞書をインポートします。",
        "operationId": "import_user_dict_words",
        "parameters": [
          {
            "description": "重複したエントリがあった場合、上書きするかどうか",
            "in": "query",
            "name": "override",
            "required": true,
            "schema": {
              "description": "重複したエントリがあった場合、上書きするかどうか",
              "title": "Override",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/UserDictWord"
                    },
                    {
                      "$ref": "#/components/schemas/UserDictWordForCompat"
                    }
                  ]
                },
                "description": "インポートするユーザー辞書のデータ",
                "title": "Import Dict Data",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "他のユーザー辞書をインポートする",
        "tags": [
          "ユーザー辞書"
        ]
      }
    },
...
    "/user_dict": {
      "get": {
        "description": "ユーザー辞書に登録されている単語の一覧を返します。\n単語の表層形 (surface) は正規化済みの物を返します。",
        "operationId": "get_user_dict_words",
        "parameters": [
          {
            "description": "複数のアクセント句を持つ単語の扱いを指定します。false の場合は API 互換性のため、最初のアクセント句の情報のみを返します。",
            "in": "query",
            "name": "enable_compound_accent",
            "required": false,
            "schema": {
              "default": false,
              "description": "複数のアクセント句を持つ単語の扱いを指定します。false の場合は API 互換性のため、最初のアクセント句の情報のみを返します。",
              "title": "Enable Compound Accent",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/UserDictWord"
                      },
                      {
                        "$ref": "#/components/schemas/UserDictWordForCompat"
                      }
                    ]
                  },
                  "title": "Response Get User Dict Words User Dict Get",
                  "type": "object"
                }
              }
            },
            "description": "単語の UUID とその詳細"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "ユーザー辞書に登録されている単語の一覧を取得する",
        "tags": [
          "ユーザー辞書"
        ]
      }
    },
    "/user_dict_word": {
      "post": {
        "description": "ユーザー辞書に単語を追加します。",
        "operationId": "add_user_dict_word",
        "parameters": [
          {
            "description": "単語の表層形",
            "in": "query",
            "name": "surface",
            "required": true,
            "schema": {
              "description": "単語の表層形",
              "items": {
                "type": "string"
              },
              "title": "Surface",
              "type": "array"
            }
          },
          {
            "description": "単語の発音（カタカナ）",
            "in": "query",
            "name": "pronunciation",
            "required": true,
            "schema": {
              "description": "単語の発音（カタカナ）",
              "items": {
                "type": "string"
              },
              "title": "Pronunciation",
              "type": "array"
            }
          },
          {
            "description": "アクセント型（音が下がる場所を指す）",
            "in": "query",
            "name": "accent_type",
            "required": true,
            "schema": {
              "description": "アクセント型（音が下がる場所を指す）",
              "items": {
                "type": "integer"
              },
              "title": "Accent Type",
              "type": "array"
            }
          },
          {
            "description": "PROPER_NOUN（固有名詞）、LOCATION_NAME（地名）、ORGANIZATION_NAME（組織・施設名）、PERSON_NAME（人名）、PERSON_FAMILY_NAME（姓）、PERSON_GIVEN_NAME（名）、COMMON_NOUN（普通名詞）、VERB（動詞）、ADJECTIVE（形容詞）、SUFFIX（語尾）のいずれか",
            "in": "query",
            "name": "word_type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/WordTypes",
              "description": "PROPER_NOUN（固有名詞）、LOCATION_NAME（地名）、ORGANIZATION_NAME（組織・施設名）、PERSON_NAME（人名）、PERSON_FAMILY_NAME（姓）、PERSON_GIVEN_NAME（名）、COMMON_NOUN（普通名詞）、VERB（動詞）、ADJECTIVE（形容詞）、SUFFIX（語尾）のいずれか"
            }
          },
          {
            "description": "単語の優先度（0から10までの整数）。数字が大きいほど優先度が高くなる。1から9までの値を指定することを推奨。",
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "description": "単語の優先度（0から10までの整数）。数字が大きいほど優先度が高くなる。1から9までの値を指定することを推奨。",
              "maximum": 10,
              "minimum": 0,
              "title": "Priority",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Add User Dict Word User Dict Word Post",
                  "type": "string"
                }
              }
            },
            "description": "追加した単語の UUID"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "ユーザー辞書に単語を追加する",
        "tags": [
          "ユーザー辞書"
        ]
      }
    },
    "/user_dict_word/{word_uuid}": {
      "delete": {
        "description": "ユーザー辞書に登録されている単語を削除します。",
        "operationId": "delete_user_dict_word",
        "parameters": [
          {
            "description": "削除する単語の UUID",
            "in": "path",
            "name": "word_uuid",
            "required": true,
            "schema": {
              "description": "削除する単語の UUID",
              "title": "Word Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "ユーザー辞書に登録されている単語を削除する",
        "tags": [
          "ユーザー辞書"
        ]
      },
      "put": {
        "description": "ユーザー辞書に登録されている単語を更新します。",
        "operationId": "rewrite_user_dict_word",
        "parameters": [
          {
            "description": "更新する単語の UUID",
            "in": "path",
            "name": "word_uuid",
            "required": true,
            "schema": {
              "description": "更新する単語の UUID",
              "title": "Word Uuid",
              "type": "string"
            }
          },
          {
            "description": "単語の表層形",
            "in": "query",
            "name": "surface",
            "required": true,
            "schema": {
              "description": "単語の表層形",
              "items": {
                "type": "string"
              },
              "title": "Surface",
              "type": "array"
            }
          },
          {
            "description": "単語の発音（カタカナ）",
            "in": "query",
            "name": "pronunciation",
            "required": true,
            "schema": {
              "description": "単語の発音（カタカナ）",
              "items": {
                "type": "string"
              },
              "title": "Pronunciation",
              "type": "array"
            }
          },
          {
            "description": "アクセント型（音が下がる場所を指す）",
            "in": "query",
            "name": "accent_type",
            "required": true,
            "schema": {
              "description": "アクセント型（音が下がる場所を指す）",
              "items": {
                "type": "integer"
              },
              "title": "Accent Type",
              "type": "array"
            }
          },
          {
            "description": "PROPER_NOUN（固有名詞）、LOCATION_NAME（地名）、ORGANIZATION_NAME（組織・施設名）、PERSON_NAME（人名）、PERSON_FAMILY_NAME（姓）、PERSON_GIVEN_NAME（名）、COMMON_NOUN（普通名詞）、VERB（動詞）、ADJECTIVE（形容詞）、SUFFIX（語尾）のいずれか",
            "in": "query",
            "name": "word_type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/WordTypes",
              "description": "PROPER_NOUN（固有名詞）、LOCATION_NAME（地名）、ORGANIZATION_NAME（組織・施設名）、PERSON_NAME（人名）、PERSON_FAMILY_NAME（姓）、PERSON_GIVEN_NAME（名）、COMMON_NOUN（普通名詞）、VERB（動詞）、ADJECTIVE（形容詞）、SUFFIX（語尾）のいずれか"
            }
          },
          {
            "description": "単語の優先度（0から10までの整数）。数字が大きいほど優先度が高くなる。1から9までの値を指定することを推奨。",
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "description": "単語の優先度（0から10までの整数）。数字が大きいほど優先度が高くなる。1から9までの値を指定することを推奨。",
              "maximum": 10,
              "minimum": 0,
              "title": "Priority",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "ユーザー辞書に登録されている単語を更新する",
        "tags": [
          "ユーザー辞書"
        ]
      }
    },
...
  }
}